feature: Separação de serviços
- [x] Criação de um serviço para o gerente
- [x] Modificação do arquivo init para receber um novo serviço
- [x] segregação de serviço de gerente e usuário comum
- [x] Criação de nova conexão do banco de dados para ambos os serviços
- [x] Script de inicialização dos serviços, arquivo run.sh para linux
